server/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
black==24.8.0
certifi==2024.8.30
click==8.1.7
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==30.3.0
fastapi==0.115.0
flake8==4.0.1
flake8-bugbear==22.1.11
flake8-pie==0.15.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.6.1
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.8.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyflakes==2.4.0
pytest==7.2.1
pytest-asyncio==0.20.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
tomli==2.0.2
types-factory-boy==0.4.1
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
uvicorn==0.31.0
